Second part, no codestyle
