+++ client_build/static/styles.css
@@ -6,11 +6,17 @@
   display: block;
   height: inherit;
 }
+
 .test-2 {
+  box-sizing: border-box;
   width: 100%;
   height: 700px;
 }
 .test-2__block {
   display: block;
   height: inherit;
+  border-radius: 10px;
+  box-shadow: 10px 10px 5px 0px rgba(0, 0, 0, 0.75);
 }
+
+/*# sourceMappingURL=styles.css.map */
