+++ client_src/less/test2.less
@@ -1,8 +1,11 @@
 .test-2 {
+	box-sizing: border-box;
 	width: 100%;
 	height: 700px;
 	&__block {
 		display: block;
 		height: inherit;
+		border-radius: 10px;
+		box-shadow: 10px 10px 5px 0px rgba(0,0,0,0.75);
 	}
 }
