+++ gulpfile.js
@@ -11,25 +11,56 @@ var argv = require('yargs').argv;
 var gulpif = require('gulp-if');
 var sourcemaps = require('gulp-sourcemaps');
 var watch = require('gulp-watch');
+var htmlmin = require('gulp-htmlmin');
+var uglify = require('gulp-uglify');
+var browserSync = require('browser-sync').create();
+var LessPluginAutoPrefix = require('less-plugin-autoprefix'),
+    autoprefix= new LessPluginAutoPrefix({browsers: ["last 2 versions"]});
+
 
 gulp.task('default', ['libs', 'build']);
-gulp.task('build', ['copy-static', 'css']);
+gulp.task('build', ['copy-static', 'css', 'js']);
+
+gulp.task('serve', function () {
+    browserSync.init({
+        server: destDir
+    });
+
+    browserSync.watch(destDir + '**/*.*').on('change', browserSync.reload);
+});
 
 gulp.task('bower', function () {
     return bower();
 });
 
+gulp.task('minify', function() {
+  return gulp.src(baseDir + '/*.html')
+    .pipe(gulpif(argv.prod, htmlmin({collapseWhitespace: true})))
+    .pipe(gulp.dest(destDir))
+});
+
 gulp.task('less', function () {
   return gulp.src(baseDir + '/less/**/*.less')
-    .pipe(less())
+    .pipe(less({plugins: [autoprefix]}))
     .pipe(gulp.dest(destDir + '/static'));
+});
+
+gulp.task('js', function () {
+  return gulp.src(baseDir + '/js/**/*.js')
+    .pipe(sourcemaps.init())
+    .pipe(concat('all.js'))
+    .pipe(gulpif(argv.prod, uglify()))
+    .pipe(sourcemaps.write('.'))
+    .pipe(gulp.dest(destDir + '/js'));
 });
 
 gulp.task('css', function () {
     return gulp.src(baseDir + '/less/**/*.less')
-        .pipe(concat('styles.css'))
+        .pipe(sourcemaps.init())
         .pipe(less())
+        .pipe(concat('styles.css'))
         .pipe(gulpif(argv.prod, cssnano()))
+        .pipe(sourcemaps.write('.'))
         .pipe(gulp.dest(destDir + '/static'));
 });
 
@@ -43,6 +74,7 @@ gulp.task('copy-static', function () {
         baseDir + '/**',
             '!' + baseDir + '/libs{,/**}',
             '!' + baseDir + '/less{,/**}',
+            '!' + baseDir + '/js{,/**}',
             '!' + baseDir + '/node_modules{,/**}',
         ])
         .pipe(gulp.dest(destDir));
