+++ package.json
@@ -17,14 +17,19 @@
   },
   "homepage": "https://github.com/mrbob89/adv-fe#readme",
   "devDependencies": {
+    "browser-sync": "^2.11.1",
     "gulp": "^3.9.1",
+    "gulp-autoprefixer": "^3.1.0",
     "gulp-bower": "0.0.13",
     "gulp-clean": "^0.3.2",
     "gulp-concat": "^2.6.0",
     "gulp-cssnano": "^2.1.1",
+    "gulp-htmlmin": "^1.3.0",
     "gulp-if": "^2.0.0",
     "gulp-sourcemaps": "^1.6.0",
+    "gulp-uglify": "^1.5.3",
     "gulp-watch": "^4.3.5",
+    "less-plugin-autoprefix": "^1.5.1",
     "run-sequence": "^1.1.5",
     "yargs": "^4.2.0"
   }
